+++ spine-unity/Assets/Spine/Editor/spine-unity/Editor/SpineAttributeDrawers.cs
@@ -591,17 +591,17 @@ namespace Spine.Unity.Editor {
 					List<Skin.SkinEntry> skinEntries = new List<Skin.SkinEntry>();
 					skin.GetAttachments(i, skinEntries);
 					foreach (Skin.SkinEntry entry in skinEntries) {
-						attachmentNames.Add(entry.PlaceholderName);
+						attachmentNames.Add(entry.Placeholder);
 					}
 
 					if (skin != defaultSkin) {
 						foreach (Skin.SkinEntry entry in skinEntries) {
-							placeholderNames.Add(entry.PlaceholderName);
+							placeholderNames.Add(entry.Placeholder);
 						}
 						skinEntries.Clear();
 						defaultSkin.GetAttachments(i, skinEntries);
 						foreach (Skin.SkinEntry entry in skinEntries) {
-							attachmentNames.Add(entry.PlaceholderName);
+							attachmentNames.Add(entry.Placeholder);
 						}
 					}